Improve voltage displaying (percents and color)

+++ app/src/main/java/com/hakito/netcar/sender/CarSender.kt
@@ -17,5 +17,5 @@ data class Sensors(
     val frontLeftRpm: Int,
     val frontRightRpm: Int,
     val rearRpm: Int,
-    val voltage: Float
+    val voltage: Int
 )
